+++ RELEASES.md
@@ -47,3 +47,13 @@ Based on:
 - [typescript v0.0.1-alpha.10] .
 ### Releases
 - [NPM v0.0.1-alpha.10] https://www.npmjs.com/package/@gbg/go-core/v/0.0.1-alpha.10 - .
+
+## 2026-02-06 00:28:07
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.706.1 (2.809.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.0.1-alpha.12] .
+### Releases
+- [NPM v0.0.1-alpha.12] https://www.npmjs.com/package/@gbg/go-core/v/0.0.1-alpha.12 - .